third_party/WebKit/LayoutTests/virtual/stable/animations-unprefixed/transitions-unprefixed.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<style>
#bg {
  transition: all 5s;
  transform: none;
}
</style>
<div id="div"></div>

<script>
div.offsetTop;

test(function() {
  div.style.transform = "translateX(100px)";
  assert_not_equals(getComputedStyle(div).transform, "translateX(100px)");
}, 'Ensure that unprefixed transitions work.');
</script>
